check position exit signal
